backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-environ==0.9.0
mysqlclient==2.0.1
psycopg2
django-cors-headers==3.5.0
pika==1.1.0
django-rest-passwordreset==1.3.0
google-api-python-client
djangorestframework-simplejwt==5.2.2
cryptography
pycryptodome
selenium 
beautifulsoup4
webdriver-manager